refactor: Split code for events and notifications
