+++ packages/server/src/schema/events/types.graphql
@@ -1,19 +1,3 @@
-type Notification {
-  id: Int!
-  unread: Boolean!
-  event: AbstractNotificationEvent
-  email: Boolean!
-  emailSent: Boolean!
-}
-
-interface AbstractNotificationEvent {
-  id: Int!
-  instance: Instance!
-  date: DateTime!
-  actor: User!
-  objectId: Int!
-}
-
 extend type Query {
   events(
     after: String
@@ -24,37 +8,17 @@ extend type Query {
     objectId: Int
     instance: Instance
   ): AbstractNotificationEventConnection!
-  notifications(
-    after: String
-    first: Int
-    emailSent: Boolean
-    email: Boolean
-    unread: Boolean
-    userId: Int
-  ): NotificationConnection!
-  notificationEvent(id: Int!): AbstractNotificationEvent
-}
-
-extend type Mutation {
-  notification: NotificationMutation!
-}
-
-type NotificationMutation {
-  setState(input: NotificationSetStateInput!): DefaultResponse
-}
-
-input NotificationSetStateInput {
-  id: [Int!]!
-  unread: Boolean!
-}
-
-type NotificationConnection {
-  nodes: [Notification!]!
-  totalCount: Int!
-  pageInfo: PageInfo!
 }
 
 type AbstractNotificationEventConnection {
   nodes: [AbstractNotificationEvent!]!
   pageInfo: PageInfo!
+}
+
+interface AbstractNotificationEvent {
+  id: Int!
+  instance: Instance!
+  date: DateTime!
+  actor: User!
+  objectId: Int!
 }
